+++ src/Profiler/ProfilerTools.cpp
@@ -50,7 +50,7 @@ namespace Debug {
          // Output the timinigs
          std::stringstream oss;
          int base = 27;
-         for(int i = 0; i < ProfilerMacro::NBREAKPOINT; i++)
+         for(int i = 0; i < static_cast<int>(ProfilerMacro::NBREAKPOINT); i++)
          {
             oss << ProfilerMacro::pointName(static_cast<ProfilerMacro::BreakPoint>(i)) << ": " << std::fixed << std::setprecision(digits) << ts(i);
